refactor: move class
 We need to move the connector, to not pull in more dependencies
